+++ packages/kysely/migrate/migrations/index.ts
@@ -77,6 +77,7 @@ import * as n77_drop_sessions from "./2025-03-14T21:58:00_drop_session";
 import * as n78_login_sessions from "./2025-03-14T22:41:00_login_sessions";
 import * as n79_drop_sessions_2 from "./2025-03-19T15:47:00_drop_sessions_2";
 import * as n80_recreate_custom_domains from "./2025-03-21T11:08:00_recreate_custom_domains";
+import * as n81_phone from "./2025-04-22T14:00:00_phone";
 
 // These need to be in alphabetic order
 export default {
@@ -159,4 +160,5 @@ export default {
   n78_login_sessions,
   n79_drop_sessions_2,
   n80_recreate_custom_domains,
+  n81_phone,
 };